Add person saving and loading functionality
