+++ data/persons.json
@@ -1,0 +1,1 @@
+[{"type":"Student","name":"aaa","age":20},{"type":"Teacher","name":"wee","age":40,"specialization":"aa"},{"type":"Student","name":"pop","age":14},{"type":"Teacher","name":"tt","age":24,"specialization":""}]
