+++ src/app.rb
@@ -1,12 +1,11 @@
-require 'json'
-require 'fileutils'
 require_relative './functions/list_books'
 require_relative './functions/list_persons'
 require_relative './functions/create_person'
 require_relative './functions/create_book'
 require_relative './functions/create_rental'
 require_relative './functions/list_rentals'
-require_relative './book'
+require_relative './functions/save_data'
+require_relative './functions/load_data'
 
 class App
   attr_reader :books, :persons, :rentals
@@ -39,25 +38,11 @@ class App
   end
 
   def save_data
-    arr = []
-    books.each do |book|
-      obj = {
-        'title' => book.title,
-        'author' => book.author
-      }
-      arr.push(obj)
-    end
-    FileUtils.mkdir_p('data')
-    File.write('./data/books.json', JSON.generate(arr))
-    puts 'Data saved successfully'
+    save_book
+    save_person
   end
 
-  def load_data
-    puts 'LOADING...'
-    file_data = JSON.parse(File.read('./data/books.json'))
-    file_data.each do |item|
-      book = Book.new(item['title'], item['author'])
-      books << book
-    end
+  def load 
+    load_data
   end
 end
