+++ src/teacher.rb
@@ -1,6 +1,7 @@
 require_relative 'person'
 
 class Teacher < Person
+  attr_reader :specialization
   def initialize(age, specialization, name = 'Unknown')
     super(age, name)
     @specialization = specialization
